Version 3.21.1
Promoted ArrayBuffer, DataView and typed arrays to non-experimental. (Chromium issue 270527)

Replaced OS::MemCopy with memcpy in typed array initialization. (Chromium issue 270642)

Moved i18n break iterator C++ code to runtime (issue 2745)

Fixed invalid out-of-bounds store in MacroAssembler::Allocate. (Chromium issue 263515)

Fixed register misuse in Allocate() on ARM. (issue 2851)

Fixed empty handle dereference in Runtime_InternalNumberFormat. (Chromium issue 275467)

Performance and stability improvements on all platforms.

git-svn-id: http://v8.googlecode.com/svn/trunk@16257 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mips/assembler-mips.cc
@@ -2203,8 +2203,7 @@ void Assembler::set_target_address_at(Address pc, Address target) {
 
   Instr instr3 = instr_at(pc + 2 * kInstrSize);
   uint32_t ipc = reinterpret_cast<uint32_t>(pc + 3 * kInstrSize);
-  bool in_range = (ipc ^ static_cast<uint32_t>(itarget) >>
-                  (kImm26Bits + kImmFieldShift)) == 0;
+  bool in_range = ((ipc ^ itarget) >> (kImm26Bits + kImmFieldShift)) == 0;
   uint32_t target_field =
       static_cast<uint32_t>(itarget & kJumpAddrMask) >> kImmFieldShift;
   bool patched_jump = false;
